reorganize file structure
